Add files via upload
The programs used to generate shapes. "Final_Program_v2.2.py" is the program from the semester 2 bachelor project and only generates squares, rectangles, and empty images. The second file "shape_generator.py" is the program written during this semester to test higher diversity in shape types and whether that was a factor in the accuracy of the Neural network model during tests.